Tue, May 30, 2023, 7:30 AM +00:00

+++ bulbs.c
@@ -28,9 +28,9 @@ int print_bulb(string message)
         for (int j = 0 ; j < 9 ; j++)
         {
             remainder[j] = ascii % 2;
-            ascii = ascii / 2;
+            ascii = (int)ascii / (int)2;
         }
     }
 
-    
+
 }
